+++ test/index.js
@@ -1,4 +1,4 @@
-var test = require('tap').test;
+var test = require('tape').test;
 var fixtures = require('./fixtures');
 var expectations = require('./expectations');
 var path = require('path');
@@ -7,7 +7,7 @@ var fs = require('fs');
 var path = require('path');
 var exec = require('child_process').exec;
 
-test('valid zipfiles', function(group) {
+test('valid zipfiles, include mandatory files', function(group) {
   Object.keys(fixtures.valid).forEach(function(k) {
     group.test(fixtures.valid[k], function(t) {
       shpFairy(fixtures.valid[k], function(err, output) {
@@ -17,7 +17,7 @@ test('valid zipfiles', function(group) {
         var base = path.basename(output, '.shp');
         var dir = path.dirname(output);
 
-        ['.shp', '.shx', '.dbf', '.prj', '.index'].forEach(function(ext) {
+        ['.shp', '.shx', '.dbf'].forEach(function(ext) {
           t.ok(fs.existsSync(path.join(dir, base + ext)), k + ': ' + ext + ' exists');
         });
 
@@ -27,7 +27,7 @@ test('valid zipfiles', function(group) {
             t.fail(k + ': unexpected ' + ext + ' file created');
           }
           var stats = fs.statSync(path.join(dir, filename));
-          t.equal(expectations.valid[k][ext], stats.size, k + ': ' + ext + ' is the correct size');
+          t.equal(stats.size, expectations.valid[k][ext], k + ': ' + ext + ' is the correct size');
         });
 
         return t.end();
@@ -35,6 +35,24 @@ test('valid zipfiles', function(group) {
     });
   });
   group.end();
+});
+
+test('valid zipfile preserves all non-mandatory files', function(t) {
+  shpFairy(fixtures.valid.everything, function(err, output) {
+    if (err) return t.end();
+
+    var base = path.basename(output, '.shp');
+    var dir = path.dirname(output);
+
+    var acceptedExtensions = ['shp', 'shx', 'dbf', 'prj', 'sbn', 'sbx', 'fbn', 'fbx', 'ain', 'aih', 'ixs', 'mxs', 'atx', 'xml', 'cpg', 'qix', 'index'];
+
+    fs.readdirSync(dir).forEach(function(filename) {
+      var ext = path.extname(filename).slice(1);
+      t.ok(acceptedExtensions.indexOf(ext) > -1, ext + ' was preserved');
+    });
+
+    t.end();
+  });
 });
 
 test('invalid zipfiles', function(group) {
@@ -94,5 +112,39 @@ test('executable script: invalid case with --quiet', function(t) {
     t.equal(stdout, '', 'empty stdout');
     t.equal(stderr, 'ZIP file did not contain a shp file\n', 'expected error message');
     t.end();
+  });
+});
+
+test('executable script: replaces single backslash as expected', function(t) {
+  var valid = [
+    path.resolve(__dirname, '..', 'bin', 'shapefile-fairy.js'),
+    fixtures.valid.backslash_single
+  ].join(' ');
+
+  exec(valid, function(err, stdout, stderr) {
+    if (err) throw err;
+    t.notOk(stderr, 'no errors logged');
+    t.notOk(/\\/.test(path.basename(stdout)), 'no backwards slashes');
+    fs.exists(stdout.slice(0, -1), function(exists) {
+      t.ok(exists, 'output exists');
+      t.end();
+    });
+  });
+});
+
+test('executable script: replaces double backslash as expected', function(t) {
+  var valid = [
+    path.resolve(__dirname, '..', 'bin', 'shapefile-fairy.js'),
+    fixtures.valid.backslash_double
+  ].join(' ');
+
+  exec(valid, function(err, stdout, stderr) {
+    if (err) throw err;
+    t.notOk(stderr, 'no errors logged');
+    t.notOk(/\\/.test(path.basename(stdout)), 'no backwards slashes');
+    fs.exists(stdout.slice(0, -1), function(exists) {
+      t.ok(exists, 'output exists');
+      t.end();
+    });
   });
 });