+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.4.0
+
+- Fix bug where the name sanitizer didn't remove single backslashes [#19](https://github.com/mapbox/shapefile-fairy/issues/19)
+- test on Node v6, remove testing for Node v0.10.x
+- add `valid.everything.zip` fixture that asserts for all possible file extensions [#1](https://github.com/mapbox/shapefile-fairy/issues/1)
+- update tests to assert for mandatory fixtures, and a separate test for non-mandatory file preservation
+
 ## 0.3.4
 
  - Updated to node-zipfile@0.5.9